Reverted file manager changes

+++ MetaMorpheus/Test/MetaDraw/MetaDrawTest.cs
@@ -1509,7 +1509,8 @@ namespace Test.MetaDraw
             string proteinDatabase = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\smalldb.fasta");
             string spectraFile = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SmallCalibratible_Yeast.mzML");
 
-            Directory.Delete(outputFolder, true);
+            if (Directory.Exists(outputFolder))
+                Directory.Delete(outputFolder, true);
             Directory.CreateDirectory(outputFolder);
 
             // run search task
